+++ lndgrpc/__init__.py
@@ -7,4 +7,4 @@ MINOR = sys.version_info[1]
 
 # only import the async client for python 3.6+
 if MAJOR == 3 and MINOR >= 6:
-    from lndgrpc.async.async_client import AsyncLNDClient
+    from lndgrpc.aio.async_client import AsyncLNDClient
